Remove float usage for optimization

+++ src/video.h
@@ -13,7 +13,7 @@
 
 bool video_init(int window_scale, float screen_x_scale, const char *quality, bool fullscreen, float opacity);
 void video_reset(void);
-bool video_step(float mhz, float steps, bool midline);
+bool video_step(uint32_t mhz, uint32_t steps, bool midline);
 bool video_update(void);
 void video_end(void);
 bool video_get_irq_out(void);
